hexdump: convert a variable type according with fmt
hexdump works uncorrectly on Rassberry Pi (raspbian wheezy):
0000000 3200000000 3400000000 3600000000 3800000000 a00000000
000000a

The problem is that the %qx format is used for printing
the (short int) variable.

Here is the output from hexdump with this patch:
0000000 3231 3433 3635 3837 0a39
000000a

Currently raspbian uses hexdump from bsdmainutils.
bsdmainutils: /usr/bin/hexdump

+++ text-utils/hexdump-display.c
@@ -93,15 +93,15 @@ print(struct hexdump_pr *pr, unsigned char *bp) {
 
 		switch(pr->bcnt) {
 		case 1:
-			printf(pr->fmt, *bp);
+			printf(pr->fmt, (unsigned long long) *bp);
 			break;
 		case 2:
 			memmove(&sval, bp, sizeof(sval));
-			printf(pr->fmt, sval);
+			printf(pr->fmt, (unsigned long long) sval);
 			break;
 		case 4:
 			memmove(&ival, bp, sizeof(ival));
-			printf(pr->fmt, ival);
+			printf(pr->fmt, (unsigned long long) ival);
 			break;
 		case 8:
 			memmove(&Lval, bp, sizeof(Lval));
@@ -130,15 +130,15 @@ print(struct hexdump_pr *pr, unsigned char *bp) {
 
 		switch(pr->bcnt) {
 		case 1:
-			printf(pr->fmt, *bp);
+			printf(pr->fmt, (unsigned long long) *bp);
 			break;
 		case 2:
 			memmove(&sval, bp, sizeof(sval));
-			printf(pr->fmt, sval);
+			printf(pr->fmt, (unsigned long long) sval);
 			break;
 		case 4:
 			memmove(&ival, bp, sizeof(ival));
-			printf(pr->fmt, ival);
+			printf(pr->fmt, (unsigned long long) ival);
 			break;
 		case 8:
 			memmove(&Lval, bp, sizeof(Lval));
